Move parseRulesAndDo to the profileparser package
Just to make the function better testable.
